refactor: Move `DeliveryCompleteEvent` to `deliverymanager.hub` package and update imports
